# server/transport/openapi/api/model_board_connect_type.go
/*
 * flop-strategy
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type BoardConnectType string

// List of BoardConnectType
const (
	BOARD_CONNECT_TYPE_DISCONNECT BoardConnectType = "BoardConnectTypeDisconnect"
	BOARD_CONNECT_TYPE_CONNECTED  BoardConnectType = "BoardConnectTypeConnected"
)